requirements.txt
asgiref==3.8.1
babel==2.13.1
Django==5.2.1
sqlparse==0.5.3
libsass==0.23.0
django-sass-processor==1.4.1
markdown==3.7
